Properly setting base's required files with Module definitions to reflect the included helpers.
