+++ lib/httpd_configmap_generator/base.rb
@@ -1,8 +1,8 @@
 require "pathname"
 require "httpd_configmap_generator/base/command"
-require "httpd_configmap_generator/base/config"
+require "httpd_configmap_generator/base/config_helper"
 require "httpd_configmap_generator/base/config_map"
-require "httpd_configmap_generator/base/file"
+require "httpd_configmap_generator/base/file_helper"
 require "httpd_configmap_generator/base/kerberos"
 require "httpd_configmap_generator/base/network"
 require "httpd_configmap_generator/base/pam"
@@ -11,6 +11,13 @@ require "httpd_configmap_generator/base/sssd"
 
 module HttpdConfigmapGenerator
   class Base
+    include Command
+    include ConfigHelper
+    include FileHelper
+    include Kerberos
+    include Network
+    include Pam
+
     APACHE_USER          = "apache".freeze
     HTTP_KEYTAB          = "/etc/http.keytab".freeze
     IPA_COMMAND          = "/usr/bin/ipa".freeze
