--- lib/httpd_configmap_generator/base/config.rb
@@ -1,13 +1,0 @@
-require "active_support"
-require "active_support/core_ext" # for Time.current
-
-module HttpdConfigmapGenerator
-  class Base
-    def config_file_backup(path)
-      if File.exist?(path)
-        timestamp = Time.current.strftime(TIMESTAMP_FORMAT)
-        FileUtils.copy(path, "#{path}.#{timestamp}")
-      end
-    end
-  end
-end
