+++ lib/httpd_configmap_generator/base/pam.rb
@@ -1,9 +1,11 @@
 module HttpdConfigmapGenerator
   class Base
-    def configure_pam
-      info_msg("Configuring PAM")
-      debug_msg("- Creating #{PAM_CONFIG}")
-      cp_template(PAM_CONFIG, template_directory)
+    module Pam
+      def configure_pam
+        info_msg("Configuring PAM")
+        debug_msg("- Creating #{PAM_CONFIG}")
+        cp_template(PAM_CONFIG, template_directory)
+      end
     end
   end
 end
